+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - new commands: 'set' and 'unset'
+- autocompletion of variable expansion, triggered by '${'
 
 ## [0.0.25] - 2019-03-01
 - support for 'PATHEXT' in Windows